refactor(ui): split the last three provider modules from their handles
DockContext, WorkbenchInboxContext and WorkbenchProjectsContext each paired a
provider component with the hook every consumer calls, so react-refresh could
never hot-replace them. Same split as the earlier context modules: the handle,
its value contract and the hook stay in a component-free `.ts` that keeps every
consumer's extensionless import valid; the provider moves to its own `.tsx`.
Only the mount points (AppShell, App) retarget.

Verbatim moves — no behavior change.
