+++ ui/src/App.tsx
@@ -40,8 +40,8 @@ import { useWindowManager } from './context/WindowManagerContext';
 import { ToastProvider } from './context/ToastProvider';
 import { ThemeProvider } from './context/ThemeProvider';
 import { VaultSandboxAppearanceBridge } from './components/VaultSandboxAppearanceBridge';
-import { WorkbenchInboxProvider } from './context/WorkbenchInboxContext';
-import { WorkbenchProjectsProvider } from './context/WorkbenchProjectsContext';
+import { WorkbenchInboxProvider } from './context/WorkbenchInboxProvider';
+import { WorkbenchProjectsProvider } from './context/WorkbenchProjectsProvider';
 import { ComposerBridgeProvider } from './context/ComposerBridgeProvider';
 import { UnsavedChangesProvider } from './context/UnsavedChangesProvider';
 import { AgentationToggle } from './components/AgentationToggle';
