+++ ui/src/context/WorkbenchProjectsProvider.tsx
@@ -1,8 +1,13 @@
-import { createContext, useCallback, useContext, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import type { ReactNode } from 'react';
 
 import { useApi } from './ApiContext';
 import type { ProjectDefaultAgent, WorkbenchProject, WorkbenchSession, WorkbenchSessionCreate } from './ApiContext';
+import {
+  WorkbenchProjectsContext,
+  type ProjectSessionsState,
+  type WorkbenchProjectsTree,
+} from './WorkbenchProjectsContext';
 import { createdReconcileMinCount } from '../lib/sessionVisibilityEvents';
 import { orderProjectSessions } from '../lib/sessionPinning';
 import { errorMessage } from '@/lib/errorMessage';
@@ -17,19 +22,6 @@ import { errorMessage } from '@/lib/errorMessage';
 const SESSIONS_PAGE_SIZE = 8;
 const RECONNECT_SESSIONS_PAGE_SIZE = 200;
 
-export interface ProjectSessionsState {
-  /** null = not loaded yet. [] = loaded-but-empty (or a first-page failure, with `error`). */
-  sessions: WorkbenchSession[] | null;
-  /** First-page (or retry) fetch in flight. */
-  loading: boolean;
-  /** Load-more (append) fetch in flight. */
-  loadingMore: boolean;
-  /** next_before_id: a string means more pages exist, null means fully loaded. */
-  cursor: string | null;
-  /** The last first-page fetch failed — any rows are kept so the user can retry. */
-  error: boolean;
-}
-
 const EMPTY_SESSIONS: ProjectSessionsState = {
   sessions: null,
   loading: false,
@@ -37,48 +29,6 @@ const EMPTY_SESSIONS: ProjectSessionsState = {
   cursor: null,
   error: false,
 };
-
-export interface WorkbenchProjectsTree {
-  projects: WorkbenchProject[] | null;
-  projectsError: string | null;
-  refreshProjects: () => Promise<void>;
-
-  sessionsOf: (projectId: string) => ProjectSessionsState;
-  expanded: ReadonlySet<string>;
-  isExpanded: (projectId: string) => boolean;
-  toggleExpanded: (projectId: string) => void;
-  loadMore: (projectId: string) => void;
-  /** Re-fetch the first page (mobile retry button / programmatic reload). */
-  reloadSessions: (projectId: string) => void;
-
-  creatingSession: (projectId: string) => boolean;
-  /** Creates a session under a project (optimistic prepend + expand) and RETURNS it;
-   *  the caller navigates (this provider is mounted outside the router). null on failure.
-   *  `overrides` lets the create surfaces pin an agent/backend; omit for the server default. */
-  createSessionForProject: (projectId: string, overrides?: Partial<WorkbenchSessionCreate>) => Promise<WorkbenchSession | null>;
-  /** Fork an existing session, prepend the new row to the source project, and return it for navigation. */
-  forkSession: (projectId: string, sessionId: string) => Promise<WorkbenchSession | null>;
-  renameProject: (projectId: string, name: string) => Promise<void>;
-  /** Persist the project's default Agent route (Project Settings) and patch the
-   *  shared cache so the sidebar + Projects page reflect it. Pass an all-null
-   *  route to clear the default back to the global default. Throws on failure
-   *  (the apiFetch layer already surfaced a toast) so the dialog can react. */
-  setProjectDefaultAgent: (
-    projectId: string,
-    route: ProjectDefaultAgent,
-    expectedAgentId: string | null,
-  ) => Promise<void>;
-  archiveProject: (projectId: string) => Promise<void>;
-  /** Throws on failure so the row's inline editor can fall back; patches title on success. */
-  renameSession: (projectId: string, sessionId: string, title: string) => Promise<void>;
-  /** Persist pin state and keep the session in the project's pinned-first order. */
-  setSessionPinned: (projectId: string, sessionId: string, pinned: boolean) => Promise<void>;
-  /** Permanently archive a session: calls the API (which reclaims its bound
-   *  tasks/watches/runs) then drops the row from the tree. Throws on failure. */
-  archiveSession: (projectId: string, sessionId: string) => Promise<void>;
-  /** After NewProjectDialog: dedup-by-id, hoist to top, expand, fetch sessions if not loaded. */
-  upsertProjectToTop: (project: WorkbenchProject) => void;
-}
 
 // Scan every project's loaded rows for a session id and apply `patch`; returns a
 // new state only when something actually changed (so unrelated consumers don't
@@ -138,8 +88,6 @@ function removeSessionRow(
 }
 
 const REORDER_ACTIVITY_EVENTS = new Set(['created', 'user_message', 'show_event']);
-
-const WorkbenchProjectsContext = createContext<WorkbenchProjectsTree | null>(null);
 
 // Single source of truth for the workbench projects/sessions tree. The desktop
 // WorkbenchSidebar (always mounted) and the mobile ProjectsPage (route) both
@@ -743,9 +691,3 @@ export const WorkbenchProjectsProvider: React.FC<{ children: ReactNode }> = ({ c
 
   return <WorkbenchProjectsContext.Provider value={value}>{children}</WorkbenchProjectsContext.Provider>;
 };
-
-export function useWorkbenchProjectsTree(): WorkbenchProjectsTree {
-  const ctx = useContext(WorkbenchProjectsContext);
-  if (!ctx) throw new Error('useWorkbenchProjectsTree must be used within a WorkbenchProjectsProvider');
-  return ctx;
-}
